+++ dizoo/board_games/chess/envs/test_chess.py
@@ -1,3 +1,8 @@
+"""
+The following code is adapted from pettingzoo:
+https://github.com/Farama-Foundation/PettingZoo/tree/master/pettingzoo/classic/chess
+"""
+
 import chess
 import chess_utils
 import numpy as np
