+++ dizoo/board_games/go/envs/test_go_env_game.py
@@ -7,6 +7,7 @@ class TestGoEnv:
 
     def test_naive(self):
         env = GoEnv(board_size=9, komi=7.5)
+        print('NOTE：actions are counted by column, such as action 9, which is the second column and the first row')
         obs, reward, done, info = env.reset()
         # env.render()
         for i in range(100):
@@ -21,19 +22,19 @@ class TestGoEnv:
             # env.render()
             if done:
                 if reward > 0:
-                    print('player 1 (human player) win')
+                    print('player 1 (black_0) win')
                 else:
                     print('draw')
                 break
 
-            """player 1"""
+            """player 2"""
             action = env.random_action()
             print('player 2 (white_0): ' + env.action_to_string(action))
             obs, reward, done, info = env.step(action)
             # env.render()
             if done:
                 if reward > 0:
-                    print('player 2 (computer player) win')
+                    print('player 2 (white_0) win')
                 else:
                     print('draw')
                 break
