+++ dizoo/board_games/go/envs/test_go_pettingzoo.py
@@ -1,9 +1,8 @@
+import pytest
 import random
 import numpy as np
-# from pettingzoo.classic import tictactoe_v3
-from dizoo.board_games.go.envs.go_env import raw_env  # pettingzoo go env
+from dizoo.board_games.go.envs.go import raw_env  # pettingzoo go env
 
-# env = tictactoe_v3.env()
 env = raw_env()
 
 
@@ -11,10 +10,15 @@ def policy(observation, agent):
     action = random.choice(np.flatnonzero(observation['action_mask']))
     return action
 
-env.reset()
-for agent in env.agent_iter():
-    observation, reward, done, info = env.last()
-    action = policy(observation, agent) if not done else None
-    env.step(action)
-    # this visualizes a single game
-    env.render()
+
+@pytest.mark.envtest
+class TestGoEnv:
+
+    def test_naive(self):
+        env.reset()
+        for agent in env.agent_iter():
+            observation, reward, done, info = env.last()
+            action = policy(observation, agent) if not done else None
+            env.step(action)
+            # this visualizes a single game
+            env.render()
